Feat:UTH-120 Create Offer Button Adjustments
* Button has been moved to display above the applicants listing on the listing page. It's now only displayed if no offers has been created
* Minor design adjustments for the button
